backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
PyMuPDF==1.24.3
python-docx==1.1.2
Pillow==10.3.0
pytesseract==0.3.10
google-generativeai==0.7.2
celery==5.4.0
redis==5.0.4
pydantic==2.7.1
httpx==0.27.0
google-cloud-vision==3.7.2
